Reset MeshParticlesMaterial, continue in separate ParticlesMaterial
